+++ routes/auth/login/index.ts
@@ -1,43 +1,8 @@
-import { auth } from "interface";
 const Router = require('express');
-const bcrypt = require('bcrypt')
-const jwt = require("jsonwebtoken");
-import sqlquery from 'database';
+import login from "middleware/auth/login";
 
 const router = Router();
 
-router.post("/", async (req: auth, res: any) => {
-  const { email, password } = req.body;
-  console.log(req.body)
-  console.log('JWT_SECRET')
-  try {
-    const sql = "SELECT * FROM users WHERE email = ?";
-    const results: any = await sqlquery(sql, [email] as never);
-
-    if (results.length === 0) {
-      return res.status(400).json({ error: "User not found" });
-    }
-
-    const user = results[0];
-    console.log(user)
-    const isMatch = await bcrypt.compare(password, user.password);
-
-    if (isMatch) {
-      const token = jwt.sign(
-        { userId: user.id, email: user.email },
-        'JWT_SECRET',
-        { expiresIn: "1h" }
-      );
-      res.json({ message: "Login successful", token });
-      const savetoken = "UPDATE users SET token = ? WHERE id = ?"
-      await sqlquery(savetoken, [token, user.id] as never)
-    } else {
-      res.status(400).json({ error: true, message: 'Password did not match' });
-    }
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: err});
-  }
-});
+router.post("/", login);
 
 export default router;
