+++ routes/auth/register/index.ts
@@ -1,24 +1,9 @@
-import { auth } from "interface/index";
 
 const Router = require('express');
-const bcrypt = require('bcrypt');
-import sqlquery from 'database';
+import register from "middleware/auth/register";
 
 const router = Router();
 
-router.post("/", async (req: auth, res: any) => {
-  const { email, password } = req.body;
-
-  try {
-    const hashedPassword = await bcrypt.hash(password, 10); // Hash password
-    const sql = "INSERT INTO users (email, password) VALUES (?, ?)";
-    await sqlquery(sql, [email, hashedPassword] as never);
-
-    res.json({ message: "User registered successfully" });
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: "Registration failed" });
-  }
-});
+router.post("/", register);
 
 export default router;
